# internal/adapters/http/docs/swagger.yaml
definitions:
  response.Error:
    properties:
      error: {}
      message:
        type: string
      request_id:
        type: string
    type: object
  response.IDResponse:
    properties:
      id:
        type: string
    type: object
  response.PaginationMeta:
    properties:
      from:
        example: 11
        type: integer
      has_next:
        example: true
        type: boolean
      has_prev:
        example: false
        type: boolean
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      to:
        example: 20
        type: integer
      total_items:
        example: 125
        type: integer
      total_pages:
        example: 13
        type: integer
    type: object
  response.Success:
    properties:
      data: {}
      message:
        type: string
      meta: {}
      request_id:
        type: string
    type: object
  taskreq.BulkDelete:
    properties:
      ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - ids
    type: object
  taskreq.BulkUpdateStatus:
    properties:
      ids:
        items:
          type: string
        minItems: 1
        type: array
      status:
        enum:
        - todo
        - in_progress
        - done
        type: string
    required:
    - ids
    - status
    type: object
  taskreq.Create:
    properties:
      description:
        maxLength: 500
        type: string
      due_date:
        type: string
      priority:
        maximum: 5
        minimum: 1
        type: integer
      tags:
        items:
          type: string
        maxItems: 10
        type: array
      title:
        minLength: 3
        type: string
    required:
    - title
    type: object
  taskreq.Update:
    properties:
      description:
        maxLength: 500
        type: string
      due_date:
        type: string
      priority:
        maximum: 5
        minimum: 1
        type: integer
      tags:
        items:
          type: string
        maxItems: 10
        type: array
      title:
        minLength: 3
        type: string
    required:
    - priority
    - title
    type: object
  taskreq.UpdateStatus:
    properties:
      status:
        enum:
        - todo
        - in_progress
        - done
        type: string
    required:
    - status
    type: object
  taskresp.BulkDelete:
    properties:
      deleted:
        items:
          type: string
        type: array
      failed:
        items:
          $ref: '#/definitions/taskresp.BulkDeleteFailure'
        type: array
    type: object
  taskresp.BulkDeleteFailure:
    properties:
      id:
        type: string
      reason:
        type: string
    type: object
  taskresp.BulkDeleteValidateError:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
  taskresp.BulkUpdateStatus:
    properties:
      failed:
        items:
          $ref: '#/definitions/taskresp.BulkUpdateStatusFail'
        type: array
      updated:
        items:
          type: string
        type: array
    type: object
  taskresp.BulkUpdateStatusFail:
    properties:
      id:
        type: string
      reason:
        type: string
    type: object
  taskresp.BulkUpdateStatusValidateError:
    properties:
      ids:
        items:
          type: string
        type: array
      status:
        items:
          type: string
        type: array
    type: object
  taskresp.CreateValidateError:
    properties:
      description:
        items:
          type: string
        type: array
      priority:
        items:
          type: string
        type: array
      tags:
        items:
          type: string
        type: array
      title:
        items:
          type: string
        type: array
    type: object
  taskresp.Detail:
    properties:
      completed:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      priority:
        type: integer
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
  taskresp.ListItem:
    properties:
      completed:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      priority:
        type: integer
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  taskresp.ListValidateError:
    properties:
      limit:
        items:
          type: string
        type: array
      page:
        items:
          type: string
        type: array
      priority:
        items:
          type: string
        type: array
      search:
        type: string
      status:
        type: string
    type: object
  taskresp.UpdateStatusValidateError:
    properties:
      status:
        items:
          type: string
        type: array
    type: object
  taskresp.UpdateValidateError:
    properties:
      description:
        items:
          type: string
        type: array
      priority:
        items:
          type: string
        type: array
      tags:
        items:
          type: string
        type: array
      title:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
  description: REST API for Go Archetype
  title: Go Archetype
  version: "1.0"
paths:
  /v1/api/tasks:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bulk DeletePublicID Tasks
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/taskreq.BulkDelete'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                data:
                  $ref: '#/definitions/taskresp.BulkDelete'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Error'
            - properties:
                errors:
                  $ref: '#/definitions/taskresp.BulkDeleteValidateError'
              type: object
      security:
      - JWTAuth: []
      summary: Bulk delete tasks
      tags:
      - tasks
    get:
      description: List tasks with pagination and filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Search keyword
        in: query
        name: search
        type: string
      - description: Entity status
        in: query
        name: status
        type: string
      - description: Entity priority
        in: query
        name: priority
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                ' meta':
                  $ref: '#/definitions/response.PaginationMeta'
                data:
                  items:
                    $ref: '#/definitions/taskresp.ListItem'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Error'
            - properties:
                errors:
                  $ref: '#/definitions/taskresp.ListValidateError'
              type: object
      security:
      - JWTAuth: []
      summary: List tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task
      parameters:
      - description: Create Entity Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/taskreq.Create'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                data:
                  $ref: '#/definitions/response.IDResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Error'
            - properties:
                errors:
                  $ref: '#/definitions/taskresp.CreateValidateError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - JWTAuth: []
      summary: Create a task
      tags:
      - tasks
  /v1/api/tasks/{public_id}:
    delete:
      parameters:
      - description: Entity Public ID
        in: path
        name: public_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - JWTAuth: []
      summary: DeletePublicID task
      tags:
      - tasks
    get:
      description: Retrieve a single task by its ID
      parameters:
      - description: Entity Public ID
        in: path
        name: public_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                data:
                  $ref: '#/definitions/taskresp.Detail'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - JWTAuth: []
      - ApiKeyAuth: []
      summary: Get task by ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Replace a task completely
      parameters:
      - description: Entity Public ID
        in: path
        name: public_id
        required: true
        type: string
      - description: Update Entity Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/taskreq.Update'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Error'
            - properties:
                errors:
                  $ref: '#/definitions/taskresp.UpdateValidateError'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - JWTAuth: []
      summary: Update task
      tags:
      - tasks
  /v1/api/tasks/{public_id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Entity Public ID
        in: path
        name: public_id
        required: true
        type: string
      - description: Update Status Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/taskreq.UpdateStatus'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Error'
            - properties:
                errors:
                  $ref: '#/definitions/taskresp.UpdateStatusValidateError'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - JWTAuth: []
      summary: Update task status
      tags:
      - tasks
  /v1/api/tasks/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Bulk Update Status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/taskreq.BulkUpdateStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                data:
                  $ref: '#/definitions/taskresp.BulkUpdateStatus'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.Error'
            - properties:
                errors:
                  $ref: '#/definitions/taskresp.BulkUpdateStatusValidateError'
              type: object
      security:
      - JWTAuth: []
      summary: Bulk update task status
      tags:
      - tasks
securityDefinitions:
  ApiKeyAuth:
    description: 'API Key Authorization header. Format: Bearer {api_key}'
    in: header
    name: Authorization
    type: apiKey
  JWTAuth:
    description: 'JWT Authorization header. Format: Bearer {jwt}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"