# internal/adapters/http/dto/request/task/bulk_update_status.go
package taskreq

type BulkUpdateStatus struct {
	IDs    []string `json:"ids" validate:"required,min=1,dive,required"`
	Status string   `json:"status" validate:"required,oneof=todo in_progress done"`
}